MonteCarloRutgers/requirements.txt
abstract-rendering==0.5.1
alabaster==0.7.6
anaconda-client==1.0.2
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astropy==1.0.4
Babel==2.1.1
bcolz==0.11.0
beautifulsoup4==4.4.1
bitarray==0.8.1
bitstring==3.1.3
blaze==0.8.3
blz==0.6.2
bokeh==0.10.0
boto==2.38.0
Bottleneck==1.0.0
certifi==14.5.14
cffi==1.2.1
clyent==0.4.0
colorama==0.3.3
conda==3.18.1
conda-build==1.18.0
conda-env==2.4.2
configobj==5.0.6
cryptography==0.9.3
Cython==0.23.3
cytoolz==0.7.3
datashape==0.4.7
datastream==0.4.11
decorator==4.0.4
Django==1.8.4
docutils==0.12
fastcache==1.0.2
Flask==0.10.1
funcsigs==0.4
glueviz==0.5.3
greenlet==0.4.9
h5py==2.5.0
idna==2.0
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.0.3
itsdangerous==0.24
jdcal==1.0
jedi==0.8.1
Jinja2==2.8
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.0.0
jupyter-console==4.0.2
jupyter-core==4.0.6
llvmlite==0.7.0+3.g1ec568f
lxml==3.4.4
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.7.1
mock==1.3.0
mongoengine==0.10.0
multipledispatch==0.4.8
nbconvert==4.0.0
nbformat==4.0.0
networkx==1.10
nltk==3.0.5
nose==1.3.7
notebook==4.0.5
numba==0.21.0
numexpr==2.4.4
numpy==1.9.3
odo==0.3.4
openpyxl==2.0.2
pandas==0.16.2
path.py==0.0.0
patsy==0.4.0
pbr==1.3.0
pep8==1.6.2
pexpect==3.3
pickleshare==0.5
Pillow==3.0.0
ply==3.8
psutil==3.2.2
ptyprocess==0.4
py==1.4.30
pyasn1==0.1.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==1.0.0
Pygments==2.0.2
pymongo==2.9
pyOpenSSL==0.15.1
pyparsing==2.0.3
pytest==2.8.1
python-dateutil==2.4.2
pytz==2015.6
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.0.1
redis==2.10.3
requests==2.7.0
rope-py3k==0.9.4.post1
runipy==0.1.3
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.16.0
simplegeneric==0.8.1
six==1.9.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.2.3
sphinx-rtd-theme==0.1.7
spyder==2.3.7
SQLAlchemy==1.0.8
statsmodels==0.6.1
sympy==0.7.6.1
tables==3.2.2
terminado==0.5
toolz==0.7.4
tornado==4.2.1
traitlets==4.0.0
ujson==1.33
Werkzeug==0.10.4
wheel==0.26.0
xlrd==0.9.4
XlsxWriter==0.7.4
xlwings==0.4.1
xlwt==1.0.0
